1.Added userList widget to follower and following list. 2. Added followUser doc in authState.
